Improvment/responsibilities split

+++ src/Web/WebAPI/Accounts.cs
@@ -9,7 +9,7 @@ public static class Accounts
 {
     public static void MapAccountApi(this GroupRouteBuilder group)
     {
-        group.MapQuery("/", (IBus bus, int? limit, int? offset, CancellationToken ct)
+        group.MapQuery("/", (IBus bus, ushort? limit, ushort? offset, CancellationToken ct)
             => ApplicationApi.GetPagedProjectionAsync<Query.ListAccounts, Projection.Account>(bus, new(limit ?? 0, offset ?? 0), ct));
 
         group.MapCommand(builder => builder.MapPost("/", (IBus bus, Command.CreateAccount command, CancellationToken ct)
@@ -21,7 +21,7 @@ public static class Accounts
         group.MapCommand(builder => builder.MapDelete("/{accountId:guid}", (IBus bus, Guid accountId, CancellationToken ct)
             => ApplicationApi.SendCommandAsync<Command.DeleteAccount>(bus, new(accountId), ct)));
 
-        group.MapQuery("/{accountId:guid}/profiles/address", (IBus bus, Guid accountId, int? limit, int? offset, CancellationToken ct)
+        group.MapQuery("/{accountId:guid}/profiles/address", (IBus bus, Guid accountId, ushort? limit, ushort? offset, CancellationToken ct)
             => ApplicationApi.GetPagedProjectionAsync<Query.ListAddresses, Projection.Address>(bus, new(accountId, limit ?? 0, offset ?? 0), ct));
 
         group.MapCommand(builder => builder.MapPut("/{accountId:guid}/profiles/billing-address", (IBus bus, Guid accountId, Dto.Address address, CancellationToken ct)
